% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameoflife.R
\name{canvasN}
\alias{canvasN}
\title{canvasN sets up a plotting area ready for subsequent plots}
\usage{
canvasN(N = 100, pts = TRUE)
}
\arguments{
\item{N}{the scale of the plotting area, default=100}

\item{pts}{a true or false term that determines whether to plot fine
points on every available integer place on the canvas.}
}
\value{
produces an empty plot and returns an empty square matrix
    of order N into which to run the dynamics; which will be plotted 
    inside the empty plot.
}
\description{
canvasN sets up an empty square plotting area of 
    dimension N. It defines an arena ont which to place plots and
    related graphics.
}
\examples{
\dontrun{
  canvasN()
  plot(c(20:80),c(20:80),type="p")
}
}
