% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrams.r
\docType{package}
\name{diagrams}
\alias{diagrams}
\alias{diagrams-package}
\title{diagrams functions to aid making sketch diagrams}
\description{
diagrams provides functions to assist in making sketch-like
    diagrams, useually for inclusion in Rmd documents. These include squares,
    oblongs, circles, and suchlike. It also has an implementation of The
    Game of Like in R. A publically available respository is available at
    https://github.com/haddonm/diagrams
}
\author{
\strong{Maintainer}: Malcolm Haddon \email{malcolm.haddon@gmail.com} (\href{https://orcid.org/0000-0001-6971-7585}{ORCID})

}
\keyword{internal}
