% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameoflife.R
\name{dolife}
\alias{dolife}
\title{dolife runs the game of life using infun for dynamics}
\usage{
dolife(N, frame, infun = dostep)
}
\arguments{
\item{N}{the size of the arena}

\item{frame}{the arena matrix}

\item{infun}{the function definning the automatic rules default
= dostep}
}
\value{
returns a revised copy of the arena
}
\description{
dolife is a wrapper function that calls an input
    function, default=dostep, which calculates the dynamics of each
    individual step for each potentially changing cell. The 
    algorithm expects to step sequentially through the 2:(N-1)
    columns and (N-1):2 rows imposing the cellular automatas
    rules, e=as expressed within infun#'
}
\examples{
\dontrun{
 num <- 100 
 arena <- glidergun(N=num)
 canvas()
 plotstep(num,arena)
 for (iter in 1:300) {
   arena <- dolife(num, arena, dostep2)
   canvas()
   plotstep(num,arena,cex=1.5,sleep=0.1)
 }
}
}
