% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameoflife.R
\name{dostep}
\alias{dostep}
\title{dostep runs a single step for a single point in the GoL}
\usage{
dostep(xj, yi, frame)
}
\arguments{
\item{xj}{the x-axis position of the point}

\item{yi}{the y-axis position of the point}

\item{frame}{the matrix holding the arena of interacting cells}
}
\value{
does the given point survive or die
}
\description{
dostep conducts the cellular automata algorithm
    developed by John Horton Conway called the Game of Life (GoL)
    on a single point in the arena defined by the xj, yi, values.
    This function is only called from within dolife but is a required
    input to dolife. In this way you can write a variant that can
    then change the implemented dynamics with the GoL (see dostep2).
    See <https://en.wikipedia.org/wiki/Conway%27s_Game_of_Life>
}
\examples{
\dontrun{
 num <- 100 
 arena <- glidergun(N=num)
 canvas()
 plotstep(num,arena)
 for (iter in 1:300) {
   arena <- dolife(num, arena, dostep)
   canvas()
   plotstep(num,arena,cex=1.5,sleep=0.1)
 }
} 
}
