% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameoflife.R
\name{dostep2}
\alias{dostep2}
\title{dostep2 runs a single step for a single point in the GoL}
\usage{
dostep2(xj, yi, frame)
}
\arguments{
\item{xj}{the x-axis position of the point}

\item{yi}{the y-axis position of the point}

\item{frame}{the matrix holding the arena of interacting cells}
}
\value{
does the given point survive or die
}
\description{
dostep2 modifies Conway's GoL cellular automaton's
    algorithm by including stochasticity on whether a new constraint
    will operate to leading to a point surviving. The single point 
    in the arena is defined by the xj, yi, values.
    This function is called from within dolife but is a required
    input to dolife. By using this function as a template you can 
    write your own variant of the survive/die algorithm and thereby
    then change the implemented dynamics with the GoL (see dostep).
}
\examples{
\dontrun{
 num <- 100 
 arena <- glidergun(N=num)
 canvas()
 plotstep(num,arena)
 for (iter in 1:300) {
   arena <- dolife(num, arena, dostep2)
   canvas()
   plotstep(num,arena,cex=1.5,sleep=0.1)
 }
} 
}
