% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameoflife.R
\name{fourblink}
\alias{fourblink}
\title{fourblink defines a matrix containing a simple blinker}
\usage{
fourblink(N = 21)
}
\arguments{
\item{N}{the size of matrix containing the blinker, default=21}
}
\value{
a matrix that contains an arrangement of points that will 
    blink each 'dolife' timestep
}
\description{
fourblink defines a matrix that contains what is known
    as a blinker. This is an arrangement of points in the arena that
    persists and appears to blink at each timestep defined by 
    'dolife'. This is a two stage blinker that repeats every two
    time steps. repeats of three and more are also possible
}
\examples{
\dontrun{
  num <- 100
  arena <- makearena(N=num)
  canvasN(101)
  mat <- fourblink()
  arena <- placematrix(arena,mat)
  plotstep(num,arena,cex=1.0)
  for (i in 1:30) {
    arena <- dolife(num, arena, dostep)
    canvasN(101)
    plotstep(num,arena,cex=1.0,sleep=0.05)
  }
}
}
