% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameoflife.R
\name{glidergun}
\alias{glidergun}
\title{glidergun defines a structure that generates repeated gliders}
\usage{
glidergun(N = 100)
}
\arguments{
\item{N}{the size of the arena over which the glidergun will operate
default=100, which is also the minimum dimension needed for an
adequater representation of the glider stream}
}
\value{
it generates an N x N matrix containing the glider gun at
    the top
}
\description{
glidergun defines a structure that will automatically
    generate a stream of gliders forever. It is best implemented 
    within a 101 x 101 canvas and needs an arena of at least 
    100 x 100. If you define a smaller arena it will default to 100.
    This is called a Godper's glider gun, I assume after the person
    who discovered it.
}
\examples{
\dontrun{
 num <- 100 
 arena <- glidergun(N=num)
 canvas()
 plotstep(num,arena)
 for (iter in 1:300) {
   arena <- dolife(num, arena, dostep)
   canvas()
   plotstep(num,arena,cex=1.5,sleep=0.1)
 }
}
}
