% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameoflife.R
\name{makeblock}
\alias{makeblock}
\title{Title}
\usage{
makeblock(r = 6, c = 6)
}
\arguments{
\item{r}{the number of rows of the solid matrix object}

\item{c}{the number of columns of the solid matrix object}
}
\value{
a matrix of shape r x c
}
\description{
Title
}
\examples{
\dontrun{ 
  num <- 100
  arena <- canvasN(100)
  mat <- makeblock(5,6)
  arena <- placematrix(arena,mat)
  plotstep(num,arena,cex=1.0)
}
}
