% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameoflife.R
\name{mixuprand}
\alias{mixuprand}
\title{mixuprand randomizes the order of digits in a number}
\usage{
mixuprand(seed)
}
\arguments{
\item{seed}{a number to be used as a random seed}
}
\value{
a number the order of whose digits have been randomized.
}
\description{
mixuprand takes a large number and randomizes the order
    of its digitsd so as to add a further layer of randomness to a 
    random seed. This is useful if using the current time to generate
    random seeds, as, obviously, repeated requests can be close in 
    time and therefore similar. This should not matter but mixuprand
    should help give the appearance of randomness.
}
\examples{
\dontrun{
  mixuprand(123456)
  mixuprand(123456) 
}
}
