% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameoflife.R
\name{pentadec}
\alias{pentadec}
\title{pentadec a fifteen step blinker}
\usage{
pentadec()
}
\value{
a matrix containing the pentadec object
}
\description{
pentadec is a blinker with a sequence of 15 steps to
    repeat. One could start the sequence at any point and it would 
    continue for ever.
}
\examples{
\dontrun{
  num <- 100
  arena <- canvasN(num)
  mat <- pentadec()
  arena <- placematrix(arena,mat)
  plotstep(num,arena,cex=1.0)
  for (i in 1:60) {
    arena <- dolife(num, arena, dostep)
    canvasN(101)
    plotstep(num,arena,cex=1.0,sleep=0.5)
  }
}  
}
