% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameoflife.R
\name{placematrix}
\alias{placematrix}
\title{placematrix puts a matrix object in the middle of an arena}
\usage{
placematrix(frame, mat)
}
\arguments{
\item{frame}{the larger arena in which to place the smaller mat}

\item{mat}{the smaller matrix to be polaced in the middle of the 
larger frame}
}
\value{
the frame object, usually the arena in GoL
}
\description{
placematrix places a matrix object inside a larger 
    matrix. This is used to place a pre-defined object inside an
    arena defined by canvasN
}
\examples{
\dontrun{
  num <- 100 
  arena <- canvasN(num)
  mat <- makeblock(9,6)
  arena <- placematrix(arena,mat)
  plotstep(num,arena,cex=1.0)
}
}
