% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameoflife.R
\name{plotcount}
\alias{plotcount}
\title{plotcount plots the population count through time}
\usage{
plotcount(cnt)
}
\arguments{
\item{cnt}{a vector of the sum of the arena at each step}
}
\value{
nothiing but it does generate a plot.
}
\description{
plotcount takes a vector of population counts, 
    identifies whether the population goes extinct of not =0, and 
    then plots the populaiton size against the number of generations
}
\examples{
\dontrun{
  num <- 100
  arena <- canvasN(num)
  mat <- pentadec()
  arena <- placematrix(arena,mat)
  plotstep(num,arena,cex=1.0)
  iter <- 100
  count <- numeric(iter)
  for (i in 1:iter) {
    arena <- dolife(num, arena, dostep)
    canvasN(num)
    plotstep(num,arena,cex=1.0,sleep=0.05)
    count[i] <- sum(arena)
  }
  plotcount(count)
}  
}
