% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameoflife.R
\name{plotstep}
\alias{plotstep}
\title{plotstep plots the arena with the current state of the GoL}
\usage{
plotstep(frame, col = 1, cex = 1, sleep = 0.2)
}
\arguments{
\item{frame}{the arena at a given time step or generation}

\item{col}{the colour of the points, default=1=black}

\item{cex}{the size of the points, default=1.0}

\item{sleep}{the interval left to enable the canvas to update and 
remain visible}
}
\value{
nothing but it does plot the arena for a given generation
}
\description{
plotstep plots a single generation within the GoL. It
    takes in the arena and plots it as points. This is designed to 
    work within RStudio's plot window
}
\examples{
\dontrun{
  num <- 100
  arena <- canvasN(num)
  mat <- pentadec()
  arena <- placematrix(arena,mat)
  plotstep(num,arena,cex=1.0)
  iter <- 100
  count <- numeric(iter)
  for (i in 1:iter) {
    arena <- dolife(num, arena, dostep)
    canvasN(num)
    plotstep(num,arena,cex=1.0,sleep=0.05)
    count[i] <- sum(arena)
  }
}
}
