% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameoflife.R
\name{random}
\alias{random}
\title{random generates a square matrix filled randomly}
\usage{
random(N = 10)
}
\arguments{
\item{N}{the size of the square matrix}
}
\value{
a square matrix with randomly filled cells
}
\description{
random generates a square matrix of order N which is 
    filled randomly using binomial random numbers giving a 50:50
    chance of any single location bing filled.
}
\examples{
\dontrun{
 set.seed(56173) # pick any random seed
 num <- 100
 arena <- canvasN(num)
 mat <- random(20)
 arena <- placematrix(arena,mat)
 plotstep(num,arena,cex=1.0)
 iter <- 100
 count <- numeric(iter)
 for (i in 1:iter) {
   arena <- dolife(num, arena, dostep)
   canvasN(num)
   plotstep(num,arena,cex=1.0,sleep=0.05)
   count[i] <- sum(arena)
 }

 plotcount(count)
}
}
