% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameoflife.R
\name{symmetric}
\alias{symmetric}
\title{symmetric is a 5x5 matrix that generates a symmetric output}
\usage{
symmetric()
}
\value{
a 5 x 5 matrix of points for use in GoL
}
\description{
symmetric is a 5 x 5 matrix that progress to form a 
    complex set of generations that have a mirrored symmetry about 
    its  vertical center. It runs for 170 generations and then 
    stabilizes into various static and three 2-stage blinkers
}
\examples{
\dontrun{
 set.seed(56173) # pick any random seed
 num <- 100
 arena <- canvasN(num)
 mat <- symmetric()
 arena <- placematrix(arena,mat)
 plotstep(num,arena,cex=1.0)
 iter <- 100
 count <- numeric(iter)
 for (i in 1:iter) {
   arena <- dolife(num, arena, dostep)
   canvasN(num)
   plotstep(num,arena,cex=1.0,sleep=0.05)
   count[i] <- sum(arena)
 }

 plotcount(count)
}
}
